+++ ch10/ex10_20_21.cpp
